+++ flexible_reports/models/validators.py
@@ -1,11 +1,7 @@
 # -*- encoding: utf-8 -*-
 from django.core.exceptions import ValidationError
 from django.template import Template
-
-try:
-    from django.utils.translation import gettext_lazy as _
-except ImportError:
-    from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 
 def TemplateValidator(value):